Interrupt-driven message fetch; merge FetchedMessages into QueueManager
Wires up InterruptedWatchdog -> FlowsManager.Interrupt -> QueueManager.Interrupt
so message fetching becomes event-driven instead of relying on a per-flow
FetchLoop background task. Each Subscribe schedules its own per-subscription
Task.Delay for timeout/suspend handling, replacing the centralized
FireTimeouts scan.

Also merges FetchedMessages into QueueManager - the split was an artifact of
the previous polling architecture and a single class better describes "the
message queue for a single flow."
